requirements.txt
streamlit
google-generativeai
python-dotenv
docx
faiss-cpu
pypdf
numpy
langchain
langchain_google_genai
langchain_community
dateparser
